Added comments to explain what code does. Fixed bug where time differences greater than 2^31 - 1 ms caused integer overflow.

+++ app/src/main/java/net/rofael/uabparkingfinder/Report.java
@@ -57,29 +57,29 @@ public class Report
     public String readableLastReportTime()
     {
         long timeDif = System.currentTimeMillis() - getReportTime();
-        int seconds = ((int) timeDif / 1000);
-        int minutes = (seconds / 60);
-        int hours = (minutes / 60);
-        int days = hours / 24;
+        long seconds = (timeDif / 1000);
+        long minutes = (seconds / 60);
+        long hours = (minutes / 60);
+        long days = hours / 24;
 
         if (days > 0)
         {
             hours = hours % 24;
-            return Integer.toString(days) + " days and " + Integer.toString(hours) + " hours ago";
+            return Long.toString(days) + " days and " + Long.toString(hours) + " hours ago";
         }
         else if (hours > 0)
         {
             minutes = minutes % 60;
-            return Integer.toString(hours) + " hours and " + Integer.toString(minutes) + " minutes ago";
+            return Long.toString(hours) + " hours and " + Long.toString(minutes) + " minutes ago";
         }
         else if (minutes > 0)
         {
             seconds = seconds % 60;
-            return Integer.toString(minutes) + " minutes and " + Integer.toString(seconds) + " seconds ago";
+            return Long.toString(minutes) + " minutes and " + Long.toString(seconds) + " seconds ago";
         }
         else
         {
-            return Integer.toString(seconds) + " seconds ago";
+            return Long.toString(seconds) + " seconds ago";
         }
